HugeGraph-81: refactor AbstractBackendStoreProvider.loadXxStore()
