+++ hugegraph-core/src/main/java/com/baidu/hugegraph/backend/store/memory/InMemoryDBStoreProvider.java
@@ -22,18 +22,12 @@ package com.baidu.hugegraph.backend.store.memory;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 
-import org.slf4j.Logger;
-
 import com.baidu.hugegraph.backend.store.AbstractBackendStoreProvider;
 import com.baidu.hugegraph.backend.store.BackendStore;
 import com.baidu.hugegraph.backend.store.memory.InMemoryDBStore.InMemoryGraphStore;
 import com.baidu.hugegraph.backend.store.memory.InMemoryDBStore.InMemorySchemaStore;
-import com.baidu.hugegraph.util.E;
-import com.baidu.hugegraph.util.Log;
 
 public class InMemoryDBStoreProvider extends AbstractBackendStoreProvider {
-
-    private static final Logger LOG = Log.logger(InMemoryDBStore.class);
 
     private static Map<String, InMemoryDBStoreProvider> providers = null;
 
@@ -53,29 +47,13 @@ public class InMemoryDBStoreProvider extends AbstractBackendStoreProvider {
     }
 
     @Override
-    public BackendStore loadSchemaStore(String name) {
-        LOG.debug("InMemoryDBStoreProvider load '{}'", name);
-
-        this.checkOpened();
-        if (!this.stores.containsKey(name)) {
-            this.stores.putIfAbsent(name, new InMemorySchemaStore(this, name));
-        }
-        BackendStore store = this.stores.get(name);
-        E.checkNotNull(store, "store");
-        return store;
+    protected BackendStore newSchemaStore(String store) {
+        return new InMemorySchemaStore(this, store);
     }
 
     @Override
-    public BackendStore loadGraphStore(String name) {
-        LOG.debug("InMemoryDBStoreProvider load '{}'", name);
-
-        this.checkOpened();
-        if (!this.stores.containsKey(name)) {
-            this.stores.putIfAbsent(name, new InMemoryGraphStore(this, name));
-        }
-        BackendStore store = this.stores.get(name);
-        E.checkNotNull(store, "store");
-        return store;
+    protected BackendStore newGraphStore(String store) {
+        return new InMemoryGraphStore(this, store);
     }
 
     @Override
